pyproject.toml
[project]
name = "dsi443-project-pipeline"
version = "0.1.0"
description = "A data pipeline that collects YouTube video metadata and uses Google Gemini AI to extract narrative structures, sentiment, and conflict dynamics for political analysis."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-genai>=1.52.0",
    "ipykernel>=7.1.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "kernel>=0.20.0",
]
